feat: introduce a memory for retry::Policy
The idea introduced, here, is to add a "memory" parameter for the retry
policy. If a long-running retryable operation (like chainsync) fails
sporadically then the downstream oura user may not actually want the
retry counter to increment. By default, the memory should be longer than
the longest possible delay. If this isn't done then max_retries will
never be encountered and retry_operation will never return. This case
(never returning) may be desirable in some cases (like for a daemon
application). But, it should not be the default case.